1. fixed date type issue when training with gym. for now, default data format is set to np.float32. added `DtypeEnv` in wrappers.py 2. added stack function for unity visual input environment.  added `stackvisual.py`. now, we can specify `stack_visual_nums` in `config.yaml:unity->env` to stack `n` visual input 3. combined `ray` and `threading` env for gym training. added `rap_wrapper.py` and `threading_wrapper.py` in `envs/wrappers/gym_wrapper` 4. optimized code structure in `Nn/networks.py`

+++ common/train/gym.py
@@ -17,7 +17,9 @@ def init_variables(env):
         newstate: [vector_obs, visual_obs]
     """
     i = 1 if env.obs_type == 'visual' else 0
-    return i, [np.array([[]] * env.n), np.array([[]] * env.n)], [np.array([[]] * env.n), np.array([[]] * env.n)]
+    return (i, 
+            [np.array([[]] * env.n, dtype=np.float32), np.array([[]] * env.n, dtype=np.float32)], 
+            [np.array([[]] * env.n, dtype=np.float32), np.array([[]] * env.n, dtype=np.float32)])
 
 def gym_train(env, model, print_func,
               begin_episode, render, render_episode,
